BSR: Move test-helper outside unit directory
