+++ test/unit/lib/services/google-sheet_test.js
@@ -1,6 +1,6 @@
 const { describe, it } = require('mocha');
 const { expect } = require('chai');
-const { sinon } = require('../../test-helper');
+const { sinon } = require('../../../test-helper');
 const axios = require('axios');
 const googleSheet = require('../../../../lib/services/google-sheet');
