+++ test/unit/lib/services/sendinblue-report_test.js
@@ -1,5 +1,5 @@
 const { describe, it } = require('mocha');
-const { sinon } = require('../../test-helper');
+const { sinon } = require('../../../test-helper');
 
 const config = require('../../../../lib/config');
 const axios = require('axios');
