+++ test/unit/lib/services/slack/commands_test.js
@@ -1,6 +1,6 @@
 const { describe, it } = require('mocha');
 const axios = require('axios');
-const { sinon } = require('../../../test-helper');
+const { sinon } = require('../../../../test-helper');
 const {
   createAndDeployPixSiteRelease,
   createAndDeployPixUI,
